Ethan Updates
Added Calculations to DeskQuote. Cost and Area are now displayed on the Index page. I was not able to figure out how to get it to display on the details page. Will this work?  Also changed "Deep" to "Depth" and added "in" and "sq in" to some spots for clarity. Finally, added drop down menus to the edit page so users can only select predetermined values.